fix: autocomplete-extended types
- Using relative path ensures that the types are included in the build

- This removes the interanl types from atom-ide module
